Change swipe up gesture settings config [1/4]
Reason for this change is to avoid us having to recompile a vendor
on devices that overlay this in the vendor partition via an auto
generated rro apk. Devices that do this include the Pixel 3 and Pixel
3 XL.

This is a problem because if the config is set to false then we lose
the ability to toggle the swipe up gesture in Settings/System/Gestures,
along with other options that we've added (full gesture mode and dt2s).

Change-Id: I170104c866eb4dfefb5174f5525e10147fb31b34
Signed-off-by: mracar07 <umutcan@umutcanacar.me>